djpace/requirement.txt
asgiref==3.5.2
async-generator==1.10
attrs==22.1.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.2
django-bulk-update-or-create==0.3.0
django-cors-headers==3.13.0
djangorestframework==3.14.0
exceptiongroup==1.0.0rc9
h11==0.14.0
idna==3.4
outcome==1.2.0
PySocks==1.7.1
python-dotenv==0.21.0
pytz==2022.4
requests==2.28.1
selenium==4.5.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.12
webdriver-manager==3.8.3
wsproto==1.2.0
